+++ src/main/java/org/example/CryptoCurrency.java
@@ -1,7 +1,7 @@
 package org.example;
 
 public class CryptoCurrency extends Asset{
-    public CryptoCurrency(String name, int code, double prix, String type_Actif, int quantite) {
+    public CryptoCurrency(String name, int code, double prix, String type_Actif, double quantite) {
         super(name, code, prix, type_Actif, quantite);
     }
