+++ src/main/java/org/example/Stock.java
@@ -2,7 +2,7 @@ package org.example;
 
 public class Stock extends Asset{
 
-    public Stock(String name, int code, double prix, String type_Actif, int quantite) {
+    public Stock(String name, int code, double prix, String type_Actif, double quantite) {
         super(name, code, prix, type_Actif, quantite);
     }
 }
